Use size_t for memory allocation sizes in debug builds. https://tracker.physiomeproject.org/show_bug.cgi?id=3811

+++ core/source/general/debug.cpp
@@ -73,7 +73,7 @@ Keeps a record of where a block of memory was allocated
 ==============================================================================*/
 {
 	char *filename_line, *type;
-	unsigned long int count,size;
+	size_t count,size;
 	void *ptr;
 	int access_count;
 }; /* struct Memory_block */
@@ -203,8 +203,8 @@ struct List_memory_data
 	int count;
 	int show_pointers;
 	int show_structures;
-	unsigned long int total;
-	unsigned long int *count_total;
+	size_t total;
+	size_t *count_total;
 }; /* struct List_memory_data */
 
 static int list_memory_block(struct Memory_block *block,
@@ -311,7 +311,7 @@ Module variables
 	reallocate */
 static int display_message_call_in_progress=0;
 static int check_memory_output_on=0,maximum_count=1;
-static unsigned long int total_allocated_memory=0;
+static size_t total_allocated_memory=0;
 static struct LIST(Memory_block) *memory_block_list =
    (struct LIST(Memory_block) *)NULL;
 #endif /* defined (MEMORY_CHECKING) */
@@ -339,7 +339,7 @@ is swallowed with the call USE_PARAMETER(dummy_void); at the start of function.
 #endif /* defined (USE_PARAMETER_ON) */
 
 #if !(defined (OPTIMISED)) || (defined (NDEBUG))
-char *allocate(unsigned long int size,const char *filename,int line, const char *type)
+char *allocate(size_t size,const char *filename,int line, const char *type)
 /*******************************************************************************
 LAST MODIFIED : 26 November 2001
 
@@ -349,7 +349,7 @@ Wrapper for allocate which keeps track of allocated memory.
 {
 	char *result;
 #if defined (MEMORY_CHECKING)
-	unsigned long int previous_total_allocated_memory;
+	size_t previous_total_allocated_memory;
 	struct Memory_block *new_block;
 #endif /* defined (MEMORY_CHECKING) */
 
@@ -443,7 +443,7 @@ Wrapper for deallocate which keeps track of allocated memory.
 ==============================================================================*/
 {
 #if defined (MEMORY_CHECKING)
-	unsigned long int previous_total_allocated_memory;
+	size_t previous_total_allocated_memory;
 	struct Memory_block *block;
 #endif /* defined (MEMORY_CHECKING) */
 
@@ -488,7 +488,7 @@ Wrapper for deallocate which keeps track of allocated memory.
 	LEAVE;
 } /* deallocate */
 
-char *reallocate(char *ptr,unsigned long int size,const char *filename,int line, const char *type)
+char *reallocate(char *ptr,size_t size,const char *filename,int line, const char *type)
 /*******************************************************************************
 LAST MODIFIED : 19 November 2001
 
@@ -498,7 +498,7 @@ Wrapper for reallocate which keeps track of allocated memory.
 {
 	char *result;
 #if defined (MEMORY_CHECKING)
-	unsigned long int previous_size,previous_total_allocated_memory;
+	size_t previous_size,previous_total_allocated_memory;
 	struct Memory_block *block,*new_block;
 #endif /* defined (MEMORY_CHECKING) */
 
@@ -656,7 +656,7 @@ actual object type and then the appropriate list function is called.
 
 	ENTER(list_memory);
 #if defined (MEMORY_CHECKING)
-	if (ALLOCATE(list_memory_data.count_total, unsigned long int, maximum_count + 1))
+	if (ALLOCATE(list_memory_data.count_total, size_t, maximum_count + 1))
 	{
 		for (i = 0 ; i < (maximum_count + 1) ; i++)
 		{
